refactor: extract update to hook
