+++ apps/web/src/pages/integrations/UpdateProviderPage.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from 'react';
-import { ChannelTypeEnum, IConfigCredentials, ICredentialsDto, IUpdateIntegrationBodyDto } from '@novu/shared';
+import { ChannelTypeEnum, IConfigCredentials, ICredentialsDto } from '@novu/shared';
 import { ActionIcon, Group, Loader, Center, Stack } from '@mantine/core';
 import { useClipboard } from '@mantine/hooks';
 import styled from '@emotion/styled';
@@ -17,10 +17,7 @@ import { IntegrationEnvironmentPill } from './components/IntegrationEnvironmentP
 import { useFetchEnvironments } from '../../hooks/useFetchEnvironments';
 import { ProviderImage } from './components/multi-provider/SelectProviderSidebar';
 import { When } from '../../components/utils/When';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { updateIntegration } from '../../api/integration';
-import { QueryKeys } from '../../api/query.keys';
-import { errorMessage, successMessage } from '../../utils/notifications';
+import { useUpdateIntegration } from '../../api/hooks/useUpdateIntegration';
 
 interface IProviderForm {
   name: string;
@@ -32,19 +29,10 @@ interface IProviderForm {
 export function UpdateProviderPage() {
   const { environments, isLoading: areEnvironmentsLoading } = useFetchEnvironments();
   const [selectedProvider, setSelectedProvider] = useState<IIntegratedProvider | null>(null);
+  const { onUpdateIntegration, isLoadingUpdate } = useUpdateIntegration(selectedProvider?.integrationId || '');
   const { providers, refetch, isLoading } = useProviders();
   const { integrationId } = useParams();
   const navigate = useNavigate();
-  const queryClient = useQueryClient();
-
-  const { mutateAsync: updateIntegrationApi, isLoading: isLoadingUpdate } = useMutation<
-    { res: string },
-    { error: string; message: string; statusCode: number },
-    {
-      id: string;
-      data: IUpdateIntegrationBodyDto;
-    }
-  >(({ id, data }) => updateIntegration(id, data));
 
   const {
     control,
@@ -82,23 +70,6 @@ export function UpdateProviderPage() {
 
   const identifierClipboard = useClipboard({ timeout: 1000 });
 
-  async function onUpdateIntegration(data) {
-    try {
-      await updateIntegrationApi({
-        id: selectedProvider?.integrationId ?? '',
-        data: { ...data },
-      });
-
-      await queryClient.refetchQueries({
-        predicate: ({ queryKey }) => queryKey.includes(QueryKeys.integrationsList),
-      });
-
-      successMessage('Instance configuration updated');
-    } catch (e: any) {
-      errorMessage(e.message || 'Unexpected error');
-    }
-  }
-
   useEffect(() => {
     if (selectedProvider && !selectedProvider?.identifier) {
       const newIdentifier = slugify(selectedProvider?.displayName, {
@@ -121,7 +92,7 @@ export function UpdateProviderPage() {
 
     setSelectedProvider(foundProvider);
     reset({
-      name: foundProvider.name,
+      name: foundProvider.name ?? foundProvider.displayName,
       identifier: foundProvider.identifier,
       credentials: foundProvider.credentials.reduce((prev, credential) => {
         prev[credential.key] = credential.value;
@@ -270,7 +241,13 @@ export function UpdateProviderPage() {
                 name={`credentials.${credential.key}`}
                 control={control}
                 defaultValue=""
-                rules={{ required: `Please enter a ${credential.displayName.toLowerCase()}` }}
+                rules={
+                  credential.required
+                    ? {
+                        required: `Please enter a ${credential.displayName.toLowerCase()}`,
+                      }
+                    : undefined
+                }
                 render={({ field }) => (
                   <IntegrationInput credential={credential} errors={errors?.credentials ?? {}} field={field} />
                 )}
